// src/app/Models/MedicalInfoDetaildById.ts
export interface MedicalInfoDetailsById {
    id?: any,
    title?: any
    description?: any
    imageName?: any
    videoName?: any
    active?: boolean
    arabicCategoryName?: any
    englishCategoryName?: any
    medicalInfoTopicCategoryDto?: any
    uploadFiles?: boolean
    uploadVideos?: boolean,
    files?: any
    filesVideo?: any
    // isUploadFile?: boolean
}